Working Version
This version works for my purposes. It filters the newest .zims in each folder at Kiwix and counts up the storage cost. If there is a maxi .zim it assumes we will download only the maxi and doesn't calculate any of the other zims in the directory.

Next I will add downloading it to ~/Desktop/storage.

I'll add a download